added downloading files

+++ pentaract/src/common/types.rs
@@ -1,2 +1,2 @@
 pub type ChatId = i64;
-pub type Position = i32;
+pub type Position = i16;
